cpumem-diagnosis: smatch errors
../../sun4/cpumem-diagnosis/cmd_memerr.c: In function 'cmd_ue_common':
../../sun4/cpumem-diagnosis/cmd_memerr.c:871: error: incompatible types when returning type 'void *' but 'cmd_evdisp_t' was expected

../../sun4/cpumem-diagnosis/cmd_cpuerr.c: In function 'cmd_xxcu_resolve':
../../sun4/cpumem-diagnosis/cmd_cpuerr.c:402: error: assignment makes integer from pointer without a cast
../../sun4/cpumem-diagnosis/cmd_cpuerr.c: In function 'cmd_xxcu_initial':
../../sun4/cpumem-diagnosis/cmd_cpuerr.c:503: error: assignment makes integer from pointer without a cast

../../sun4/cpumem-diagnosis/cmd_cpu.c: In function 'cmd_trw_alloc':
../../sun4/cpumem-diagnosis/cmd_cpu.c:1316: error: comparison between pointer and integer
/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: ../../sun4/cpumem-diagnosis/cmd_cpu.c:99 cmd_cpu_type2name() error: buffer overflow 'cpu_names' 12 <= 12
/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: ../../sun4/cpumem-diagnosis/cmd_cpu.c:99 cmd_cpu_type2name() error: buffer overflow 'cpu_names' 12 <= 12

/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: cmd_Lxcacheerr.c:535 gen_data_for_ecc() warn: should '((state[3] >> i) & 1) << bit_position' be a 64 bit type?
/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: cmd_Lxcacheerr.c:561 gen_data_for_ecc() warn: should '((state[2] >> i) & 1) << bit_position' be a 64 bit type?

+++ usr/src/cmd/fm/modules/sun4u/cpumem-diagnosis/cmd_Lxcacheerr.c
@@ -532,7 +532,8 @@ gen_data_for_ecc(uint64_t *tag_data, uint64_t *data_for_ecc_gen,
 				    (((state[1] >> i) & 1) << bit_position);
 				bit_position++;
 				data_for_ecc_gen[1] |=
-				    (((state[3] >> i) & 1) << bit_position);
+				    (((uint64_t)(state[3] >> i) & 1) <<
+				    bit_position);
 				bit_position++;
 			}
 			/*
@@ -558,7 +559,8 @@ gen_data_for_ecc(uint64_t *tag_data, uint64_t *data_for_ecc_gen,
 				    (((state[0] >> i) & 1) << bit_position);
 				bit_position++;
 				data_for_ecc_gen[0] |=
-				    (((state[2] >> i) & 1) << bit_position);
+				    (((uint64_t)(state[2] >> i) & 1) <<
+				    bit_position);
 				bit_position++;
 			}
 			/*
